lms_api/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
mysqlclient==2.0.3
packaging==23.1
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.3.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
